7-Zip Version 4.10 Beta
-  Codecs now use new interface: ICompressSetDecoderProperties2

+++ 7zip/Archive/GZip/GZipHandler.cpp
@@ -143,7 +143,7 @@ STDMETHODIMP CHandler::GetProperty(UInt32 index, PROPID propID,  PROPVARIANT *va
       FILETIME utcTime;
       if (m_Item.Time != 0)
       {
-        NTime::UnixTimeToFileTime(m_Item.Time, utcTime);
+        NTime::UnixTimeToFileTime((UInt32)m_Item.Time, utcTime);
         propVariant = utcTime;
       }
       else
